* Some changes for MSVC.
git-svn-id: https://qucs.svn.sourceforge.net/svnroot/qucs/trunk@1286 b5b04e8c-4942-46c9-ab4f-83783d557d1c

+++ qucs/dialogs/qucssettingsdialog.cpp
@@ -373,7 +373,7 @@ void QucsSettingsDialog::slotApply()
     QucsSettings.Editor = editorEdit->text();
     changed = true;
   }
-  if(QucsSettings.NodeWiring != checkWiring->isChecked()) {
+  if(QucsSettings.NodeWiring != (unsigned)checkWiring->isChecked()) {
     QucsSettings.NodeWiring = checkWiring->isChecked();
     changed = true;
   }
